+++ src/app/myPage/orderDelivery/webAsideMenu.tsx
@@ -12,13 +12,10 @@ export default function WebAsideMenu() {
       <h2 className={orderDelivery_styles.web_order_delivery_list_tit}>쇼핑 정보</h2>
       <ul className={orderDelivery_styles.web_order_delivery_list_text}>
         <li>
-          <Link href="#">• 주문/배송 조회</Link>
+          <Link href="/myPage/orderDelivery">• 주문/배송 조회</Link>
         </li>
         <li>
-          <Link href="#">• 취소/반품/교환 내역</Link>
-        </li>
-        <li>
-          <Link href="#">• 찜 리스트</Link>
+          <Link href="#/myPage/wishlist">• 찜 리스트</Link>
         </li>
       </ul>
 
@@ -26,10 +23,10 @@ export default function WebAsideMenu() {
       <h2 className={orderDelivery_styles.web_order_delivery_list_tit}>혜택 관리</h2>
       <ul className={orderDelivery_styles.web_order_delivery_list_text}>
         <li>
-          <Link href="#">• 스탬프</Link>
+          <Link href="/myPage/stamp">• 스탬프</Link>
         </li>
         <li>
-          <Link href="#">• 당첨 내역</Link>
+          <Link href="/myPage/rewards">• 당첨 내역</Link>
         </li>
       </ul>
 
@@ -37,16 +34,16 @@ export default function WebAsideMenu() {
       <h2 className={orderDelivery_styles.web_order_delivery_list_tit}>회원 정보</h2>
       <ul className={orderDelivery_styles.web_order_delivery_list_text}>
         <li>
-          <Link href="#">• 회원 정보 수정</Link>
+          <Link href="/myPage/authCheck/editUserInfo">• 회원 정보 수정</Link>
         </li>
         <li>
-          <Link href="#">• 마케팅 수신 동의</Link>
+          <Link href="/myPage/marketingAgreement">• 마케팅 수신 동의</Link>
         </li>
         <li>
-          <Link href="#">• 회원 탈퇴</Link>
+          <Link href="/myPage/accountDelete">• 회원 탈퇴</Link>
         </li>
         <li>
-          <Link href="#">• 배송지 관리</Link>
+          <Link href="/addShippingAddress">• 배송지 관리</Link>
         </li>
       </ul>
     </aside>
