+++ src/app/myPage/stamp/StampWebMainContent.tsx
