+++ src/app/myPage/stamp/StampWebSideContent.tsx
