+++ src/app/myPage/stamp/page.tsx
@@ -1,9 +1,7 @@
 import stamp_page_styles from '@/styles/stamp/stamp.module.css';
-import StampMobileUserInfo from '@/app/stamp/stampMobileUserInfo';
-import StampMobileBoard from '@/app/stamp/stampMobileBoard';
-import StampWebView from '@/app/stamp/stampWebView';
-
-
+import StampMobileUserInfo from '@/app/myPage/stamp/stampMobileUserInfo';
+import StampMobileBoard from '@/app/myPage/stamp/stampMobileBoard';
+import StampWebView from '@/app/myPage/stamp/stampWebView';
 
 export default function StampPage() {
   return (
@@ -20,8 +18,6 @@ export default function StampPage() {
 
       {/* 웹뷰 영역 */}
       <StampWebView />
-
-     
     </>
   );
 }
