+++ src/app/myPage/stamp/stampMobileUserInfo.tsx
